Give meshes world coordinates
The world coordinates are appended at render time to the mesh's vertex
coordinates.

So we can have prop meshes where each instance of a prop type shares
the vertices of that type via a pointer, but can be assigned a unique
position in the world via the non-shared world coordinates.